+++ packages/core/src/types/external/OrgSchema/DefinedTerm.ts
