+++ packages/core/src/types/external/OrgSchema/Thing.ts
