+++ packages/core/src/types/external/OrgSchema/VoteAction.spec.ts
