+++ packages/core/src/types/external/OrgSchema/VoteAction.ts
