+++ packages/core/src/types/external/OrgSchema/private/orgSchemaProperties.spec.ts
