+++ packages/core/src/types/external/OrgSchema/private/orgSchemaProperty.ts
