+++ packages/core/src/utils/getOrgSchemaMessage.ts
@@ -1,6 +1,6 @@
 import { type WebChatActivity } from '../types/WebChatActivity';
-import { parseCreativeWork, type CreativeWork } from '../types/external/OrgSchema2/CreativeWork';
-import { parseThing } from '../types/external/OrgSchema2/Thing';
+import { parseCreativeWork, type CreativeWork } from '../types/external/OrgSchema/CreativeWork';
+import { parseThing } from '../types/external/OrgSchema/Thing';
 
 export default function getOrgSchemaMessage(
   graph: readonly WebChatActivity['entities'][0][]
